% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnvDATA.R
\name{create_data_grid}
\alias{create_data_grid}
\title{Create gridded data for querying Movebank Env-Data}
\usage{
create_data_grid(lats, lons, times)
}
\arguments{
\item{times}{vector of POSIXct times. See example code for how to generate them.}

\item{{lons, }{lats} vectors of longitude and latitude for the desired grid.
height.above.ellipsoid will be assigned as NA.}
}
\value{
data frame with correct formatting
}
\description{
Takes a vector of latitudes and longitudes (typically evenly spaced) and a vector of times and generates the strictly formatted data frame needed to upload to Env-Data to obtain covariates for movebank. The location vector can be single location, and time vector can be single time.
}
\examples{
 library(lubridate)
 lats <- seq(38.8, 39.0, length = 40)
 lons <- seq(-77.12, -76.91, length = 40)
 # times need to be POSIXct, not Date
 start <- ymd_hms("2014-01-01 0:00:00")
 finish <- ymd_hms("2014-12-31 0:00:00")
 timestep <- ddays(20)
 times <- seq(start, finish, by = timestep)
 test_grid <- create_data_grid(lats, lons, times)

}
