% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnvDATA.R
\name{create_request}
\alias{create_request}
\title{Creates Env-DATA XML query}
\usage{
create_request(
  request_name = "MyRequest",
  type_name,
  variable_name,
  interpolation_type,
  dist_func_spatial = "geodetic",
  email = "me@my.place",
  save_to_path
)
}
\arguments{
\item{request_name}{Name for request}

\item{type_name}{Data type requested (e.g. "modis-land/MOD13A1.005"). Go to \link{http://www.bioinfo.mpg.de/orn-gateway/services2.jsp} to see list. To be entered exactly as seen with no spaces.}

\item{variable_name}{Variable name (e.g. "500m 16 days EVI").  See list at link above.}

\item{interpolation_type}{One of "bilinear", "inverse-distance-weighted", "nearest-neighbor".  See \link{https://www.movebank.org/node/6400} for details}

\item{dist_func_spatial}{function used for distance computation ... "geodetic" default generally fine.}

\item{email}{email address for notification}

\item{savefile}{whether or not to save a file (for actual use, must be TRUE)}

\item{fileout}{a file name - will be prompted if left empty}
}
\value{
a \code\link{XMLNode} object which enters the request infromation onto the online EnvData form.
}
\description{
Creates Env-DATA XML query
}
\examples{
request_name <- "EVI_request"
type_name <- "modis-land/MOD13A1.005"
variable_name <- "500m 16 days EVI"
interpolation_type <- "bilinear"
create_request("MyRequest", type_name, variable_name, interpolation_type, save_to_path = "testfile.xml")
}
\seealso{
createEnvDataGrid, uploadEnvDataRequest
}
