% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnvDATA.R
\name{download_result}
\alias{download_result}
\title{Download and return annotated data from Env-DATA}
\usage{
download_result(request_status, file_path = NULL)
}
\arguments{
\item{request_status}{the request_status returned by \code{\link{upload_request}}}

\item{file_path}{filename (and location) to save annotated data to.}
}
\value{
status of download
}
\description{
Download and return annotated data from Env-DATA
}
