% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnvDATA.R
\name{submit_track_request}
\alias{submit_track_request}
\alias{submit_grid_request}
\title{Submit EnvData Annotation Request}
\usage{
submit_track_request(track_df, para_list, user, password)

submit_grid_request(para_list, user, password)
}
\arguments{
\item{track_df}{A data.frame that in EnvData track request format, for
example converted from \code{to_track_df}}

\item{para_list}{Parameter list}

\item{user}{Movebank EnvData backend login (not the same as a MoveBank
account), which you have to contact the Movebank to obtain}

\item{password}{Movebank EnvData backend password}
}
\value{
The status url, which can be feed to \code{download_result}
}
\description{
Submit EnvData Annotation Request
}
\section{Functions}{
\itemize{
\item \code{submit_track_request}: Submit Track Annotation Request

\item \code{submit_grid_request}: Submit Grid Annotation Request
}}

