% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnvDATA.R
\name{upload_request}
\alias{upload_request}
\title{Upload an Env-Data request for Annotation}
\usage{
upload_request(
  csv_path,
  xml_path,
  login = NULL,
  password = NULL,
  link_path = "annotation_status_link.url"
)
}
\arguments{
\item{csv_path}{Location and name of \code{.csv} file to annotate, either a MoveBank track or file generated by \code{\link{createEnvDataGrid}}}

\item{xml_path}{Location and name of \code{.xml} file that contains the request instructions - almost certainly generated by \code{\link{createEnvDataRequest}}}

\item{link_path}{name of link to access status and final annotated data}

\item{{login, password}}{login and password for EnvData (NOT the same as MoveBank login and password)}
}
\value{
The returned status of request
}
\description{
Upload an Env-Data request for Annotation
}
\details{
The function will fill out the data annotation form on the EnvData annotation tool at \link{http://www.bioinfo.mpg.de/orn-gateway/services2.jsp}.  It will also create a link to the status report of the annotation request, and will attempt to automatically open that page.
}
