% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnvDATA.R
\name{create_envdata_request}
\alias{create_envdata_request}
\title{Title}
\usage{
create_envdata_request(
  data_type = c("NDVI", "EVI"),
  ext,
  buffer = 0,
  res_space = 1,
  res_time = 1,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
